pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ha-wattpilot"
version = "0.5.0"
description = "Home Assistant integration for Fronius Wattpilot EV chargers"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "wattpilot>=0.2",
    "pyyaml>=5.3.0",
    "importlib_metadata>=4.0.0",
    "aiofiles>=23.2.1",
    "packaging>=24.0",
]

[project.optional-dependencies]
dev = [
    "homeassistant>=2025.6.3",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-homeassistant-custom-component>=0.13.0",
    "ruff>=0.14.4",
    "pre-commit>=4.0.0",
    "codespell>=2.4.0",
    "yamllint>=1.37.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-homeassistant-custom-component>=0.13.0",
]

[project.urls]
"Homepage" = "https://github.com/ruaan-deysel/ha-wattpilot"
"Bug Tracker" = "https://github.com/ruaan-deysel/ha-wattpilot/issues"
"Documentation" = "https://github.com/ruaan-deysel/ha-wattpilot#readme"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["custom_components/wattpilot"]
omit = [
    "custom_components/wattpilot/wattpilot/*",
    "tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
fail_under = 90
